refactor and fix the import error during the build process of llama model
